requirements.txt
aiogram==2.25
aiohttp==3.8.1
multidict==6.0.2
yarl==1.7.2
python-dotenv
httpx
requests
beautifulsoup4